Oops, add dependency on rescript-mode
Without this you need to install rescript-mode by hand when setting up vanilla
emacs, contrary to my README.md in rescript-mode.

I think I previously said somewhere that lsp-rescript does not depend on
rescript-mode, but in fact lsp-rescript does configure

(add-to-list 'lsp-language-id-configuration '(rescript-mode . "rescript"))

I guess it's not really required to put that line in this package (it could just
go in the rescript-mode README.md instead) but it seems to be conventional (on
the basis of checking exactly one similar LSP configuration package,
lsp-julia...).

Nothing would stop you from changing lsp-language-id-configuration to tell LSP
to use some other major mode with this same LSP client (for what that's worth:
the only non-boilerplate thing in the project is to make the "Start a build?"
prompt work, so this project is mostly useful as working installable
boilerplate).